source/packages/@ada/infra/src/connectors/sources/google_bigquery/docker-image/requirements.txt
cachetools==5.0.0
certifi==2021.10.8
charset-normalizer==2.0.12
fastavro==1.4.10
google-api-core==2.7.1
google-auth==2.6.3
google-cloud-bigquery==2.34.3
google-cloud-bigquery-storage==2.13.1
google-cloud-core==2.2.3
google-crc32c==1.3.0
google-resumable-media==2.3.2
googleapis-common-protos==1.56.0
grpcio==1.44.0
grpcio-status==1.44.0
idna==3.3
jmespath==1.0.0
numpy==1.21.5
packaging==21.3
proto-plus==1.20.3
protobuf==3.20.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyparsing==3.0.7
python-dateutil==2.8.2
pytz==2022.1
requests==2.27.1
rsa==4.8
s3transfer==0.5.2
six==1.16.0
smart-open==5.2.1
urllib3==1.26.9
db-dtypes==1.0.1
